MNT: update for thin/thick sample layout + split up layout and plans
